Make custom operator new size params size_t for x64 (x64 fix)
The placement operator new / new[] overloads for the array and heap
classes declared their first (size) parameter as uint. The C++ standard
requires it to be size_t; on Win32 uint == size_t so it compiled, but on
x64 size_t is 64-bit, so every one is ill-formed (C2821 "first formal
parameter to 'operator new' must be 'size_t'"). Because Array.hpp is
included nearly everywhere, that single defect cascaded into ~4000 of the
x64 build's errors and corrupted parsing on into CRT headers.

Change the first parameter to size_t in all of them:
- Array.hpp    operator new(size_t, CMArrayT<T,A>&)
- SparseArray.hpp operator new(size_t, CSparseArray<T>&)
- FastHeap.hpp operator new / new[] (size_t, CDArray<T>& / CFastHeap&),
               both the friend declarations and the definitions.

Alignment parameters stay uint; the internal CFastHeap::pvAllocate stays
uint (operator new narrows into it - a warning, not an error - which a
later pass can widen).

No-op on Win32 (size_t == uint): default VER_ASM=TRUE build unchanged,
green, smoke-launched. Confirmed against the diagnostic x64 build: the
Array.hpp/FastHeap.hpp cascade is gone and the total error count drops
from 10263 to 6385. Found by the x64 diagnostic build, not the asm survey
(this is not asm at all). Next x64 root cause: the CFPUState FPU-control
__asm in FloatDef.hpp.

+++ jp2_pc/Source/Lib/Std/Array.hpp
@@ -717,7 +717,7 @@ public:
 	}
 
 	//******************************************************************************************
-	inline friend void* operator new(uint u_size, CMArrayT<T, A>& ma)
+	inline friend void* operator new(size_t u_size, CMArrayT<T, A>& ma)
 	{
 		// Allocate one element at a time only, and only of type T.
 		// Unfortunately, this is enforced only by checking u_size.
